Adapted embeddings to use word vectors to compute similarity in the context of extendParsing()
